# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Show Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { Show } from "./show";

export interface DashboardSummary {
  totalShows: number;
  activeShows: number;
  archivedShows: number;
  totalOverdue: number;
  nextUpShow?: Show | null;
}