# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Show Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: number;
  showId: number;
  title: string;
  url: string;
  createdAt: Date;
}