# lib/api-zod/src/generated/types/show.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Show Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface Show {
  id: number;
  name: string;
  moveInDate: Date;
  venue?: string | null;
  advanceWarehouseDate?: Date | null;
  discountDeadline?: Date | null;
  onlineOrderDeadline?: Date | null;
  showStart?: Date | null;
  dismantleDate?: Date | null;
  createdAt: Date;
  taskCount?: number;
  completedTaskCount?: number;
  eblastCount?: number;
  sentEblastCount?: number;
  overdueCount?: number;
  lastEblastName?: string | null;
  lastEblastDate?: string | null;
  exhibitorKitSent?: boolean;
  exhibitorKitDate?: string | null;
  fireMarshalStatus?: string | null;
  fireMarshalDate?: string | null;
  idSignStatus?: string | null;
}