# lib/api-zod/src/generated/types/showWithItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Show Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { Eblast } from "./eblast";
import type { Link } from "./link";
import type { Show } from "./show";
import type { Task } from "./task";

export type ShowWithItems = Show & {
  tasks?: Task[];
  eblasts?: Eblast[];
  links?: Link[];
};