# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Show Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  showId: number;
  name: string;
  category?: string | null;
  dueDate?: Date | null;
  dueDateRule?: string | null;
  completed: boolean;
  completedAt?: Date | null;
  notes?: string | null;
  createdAt: Date;
}